requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==3.0.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dotenv==1.0.1
pywin32==306
SQLAlchemy==2.0.36
tk==0.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
